Fix all/multi day events.  Closes #21
If the "date" atribute of "start" object is populated rather than the
"dateTime", grab the value of "date" in both "start" and "end" objects
so they can get formatted as all day and/or multi day events
respectfully.  Also with this change "baseURL" is now a constant, as its
value shouldn't change during execution.

Signed-off-by: Adam Jimerson <vendion@gmail.com>